+++ network/route_darwin.go
@@ -5,12 +5,12 @@ import (
 	"github.com/HunanTV/eru-agent/logs"
 )
 
-func AddVLan(vethName, ips, cid string) bool {
+func SetDefaultRoute(cid, gateway string) bool {
 	_, err := g.Docker.InspectContainer(cid)
 	if err != nil {
 		logs.Info("VLanSetter inspect docker failed", err)
 		return false
 	}
-	logs.Info("Add VLAN device success", cid, vethName)
+	logs.Info("Set default route success", cid, gateway)
 	return true
 }
