+++ network/vlan_linux.go
@@ -2,9 +2,7 @@ package network
 
 import (
 	"net"
-	"os/exec"
 	"runtime"
-	"strconv"
 
 	"github.com/HunanTV/eru-agent/g"
 	"github.com/HunanTV/eru-agent/logs"
@@ -47,7 +45,7 @@ func setUpVLan(cid, vethName, ips string, pid int) bool {
 
 	netlink.NetworkLinkAddIp(ifc, ip, ipNet)
 	netlink.NetworkLinkUp(ifc)
-	logs.Info("Add vlan device success", cid)
+	logs.Info("Add vlan device success", cid[:12])
 	return true
 }
 
@@ -55,7 +53,7 @@ func AddVLan(vethName, ips, cid string) bool {
 	lock.Lock()
 	defer lock.Unlock()
 	device, _ := Devices.Get(cid, 0)
-	logs.Info("Add new VLan to", vethName, cid)
+	logs.Info("Add new VLan to", vethName, cid[:12])
 
 	container, err := g.Docker.InspectContainer(cid)
 	if err != nil {
@@ -76,32 +74,4 @@ func AddVLan(vethName, ips, cid string) bool {
 	}
 
 	return setUpVLan(cid, vethName, ips, container.State.Pid)
-}
-
-func SetDefaultRoute(cid, gateway string) bool {
-	lock.Lock()
-	defer lock.Unlock()
-	logs.Info("Set", cid[:12], "default route", gateway)
-
-	container, err := g.Docker.InspectContainer(cid)
-	if err != nil {
-		logs.Info("RouteSetter inspect docker failed", err)
-		return false
-	}
-
-	pid := strconv.Itoa(container.State.Pid)
-	cmd := exec.Command("nsenter", "-t", pid, "-n", "route", "del", "default")
-	if err := cmd.Run(); err != nil {
-		logs.Info("Clean default route failed", err)
-		return false
-	}
-
-	cmd = exec.Command("nsenter", "-t", pid, "-n", "route", "add", "default", "gw", gateway)
-	if err := cmd.Run(); err != nil {
-		logs.Info("RouteSetter set default route failed", err)
-		return false
-	}
-
-	logs.Info("Set default route success", cid[:12], gateway)
-	return true
 }
